Attempted #1 to Create Wordle
Selects word randomly from text file valid solutions, takes input from the user comparing inputted letters (as a char array) to the solution, highlighting inputted characters in red if they are not in the solution, in yellow if they are in the solution but not in the correct spot, or in green if they are in the solution and in the correct spot.

Currently cannot differentiate between when there is more than one letter in the correct spot, or when there are two. That is to say, if inputting a word with the same letter twice (both in the wrong spot), both will appear yellow even if there is only one of said letter in the solution.